Create prolog.pl
In Prolog, you don't typically use print statements as in imperative languages. Instead, you work with facts and queries. However, you can use a built-in predicate like write/1 to display text. 

In the Prolog code, write/1 is used to print the text, and nl is used to print a newline character for formatting.